refactor(Slug): break Slug callout into composable components
