+++ packages/react/src/index.js
@@ -298,4 +298,7 @@ export { DefinitionTooltip } from './components/Tooltip/DefinitionTooltip';
 export { GlobalTheme, Theme, useTheme } from './components/Theme';
 export { usePrefix } from './internal/usePrefix';
 export { useIdPrefix } from './internal/useIdPrefix';
-export { Slug as unstable__Slug } from './components/Slug';
+export {
+  Slug as unstable__Slug,
+  SlugContent as unstable__SlugContent,
+} from './components/Slug';
