+++ packages/styles/scss/components/slug/_slug.scss
@@ -309,7 +309,7 @@ $sizes: (
   }
 
   // Slug callout styles
-  .#{$prefix}--slug.#{$prefix}--slug--enabled .#{$prefix}--popover-content {
+  .#{$prefix}--slug.#{$prefix}--slug--enabled .#{$prefix}--slug-content {
     border: 1px solid $border-subtle;
     border-radius: 16px;
     // 84px seems to make this fully opaque?
@@ -339,8 +339,24 @@ $sizes: (
   }
 
   .#{$prefix}--slug.#{$prefix}--slug--enabled .#{$prefix}--toggletip-content {
+    max-inline-size: initial;
     padding-block-end: convert.to-rem(80px);
     padding-block-start: convert.to-rem(32px);
     padding-inline: convert.to-rem(32px);
+  }
+
+  .#{$prefix}--slug.#{$prefix}--slug--enabled .#{$prefix}--slug-actions {
+    position: absolute;
+    justify-content: flex-end;
+    column-gap: 0;
+    inset-block-end: 0;
+    inset-inline-end: 0;
+  }
+
+  .#{$prefix}--slug.#{$prefix}--slug--enabled
+    .#{$prefix}--slug-actions
+    .#{$prefix}--btn--primary {
+    order: 1;
+    border-end-end-radius: convert.to-rem(16px);
   }
 }
